Move attributes to own submodule
